+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Other
 -----
 * Cassandra 2.0 support removal (PYTHON-716)
 * cqlengine: disallow Counter create, save operations (PYTHON-497)
+* cqlengine: remove the negative indices slicing support in ModelQuerySet (PYTHON-875)
 
 3.12.0
 ======